ansible/roles/flannel: allow flannel traffic through
This patch adds two iptables rules. The first one is:

    -I INPUT 1 -p udp --dport 8472 -j ACCEPT

This rule ensures that the vxlan port is open, which is needed by
flannel when using the vxlan backend. An equivalent rule is added when
firewalld is present.

The second one is:

    -I FORWARD 1 -i flannel.1 -o docker0 -j ACCEPT

This is required so that packets decapsulated by flannel can pass
through to the docker bridge. We use direct rules here when firewalld is
present.