ansible/roles/tasks: allow kube-proxy redirected traffic
The kube-proxy adds REDIRECT rules to the PREROUTING chain of the nat
table so that containers are able to access kubernetes services with IPs
in the fictitious service subnet. However, when these rules are
triggered, the resulting packets will be addressed to the docker bridge
and thus would go through the INPUT chain of the filter table, where
they would get blocked.

This patch adds a rule to allow these packets by letting through packets
originating from the docker bridge.