Blunderbuss: don't panic if less than 4 lines changed
Basically the code does:
```
    log := math.Log10($NUM_LINES_CHANGED)
    logPlus := log + .5
    fileWeight := int64(logPlus)
```
So lets assume that $NUM_LINES_CHANGED == 3.
```
    log        == 0.4771212547196624
    logPlus    == 0.9771212547196624
    fileWeight == 0
```
This means that further on weightSum (and other things) will be calculated by
a multiple by fileWeight and thus will get the value of 0. Both

`100.0*float64(potentialOwners[cur])/float64(weightSum)`
  and
`selection := rand.Int63n(weightSum)`

Will cause a panic if weightSum == 0

I solve this by adding 1 to the fileWeight after it is truncated to
int64. So we can never get a value of 0.

We also make sure the *file.Changes != 0 when setting the fileWeight.
Otherwise we do log10(0), which doesn't panic, but gives a total junk
answer, if you ask me...