mungegithub: reduce time between api call and user of issue information
We were getting ALL issues from github and then processing them 1 by 1.
We realized that it was taking a long time to process issues and so by
the time we got to the 200th open PR it could be way out of date.  So
we switched to getting a list of all issues and then right before we
used the issue we would get both the issue and pr. This meant it was as
absolutely fresh as possible.

This PR instead pulls one "page" of issues and processes that page
before getting another "page". The PR will be absolutely fresh, but the
issue may be delayed up to one "page" of issues. So we should be working
on data less than about 1-2 minutes old.

It also simplifies some code and makes the ForEachPR and ForEachIssue
more sensible.