Rewrite/simplify e2e/jenkin
The e2e stuff talked about jenkins like it cared about the given PR.
It doesn't care.  The only thing that mattered was if the e2e on jenkins
was stable or not.

Keep the PR stuff out in the submit-queue and only do the jenkins e2e
stuff in the e2e code.