Recheck mergeability if github error about branch changed on merge
We found in the logs:
 Base branch was modified. Review and try the merge again.
On some PRs that were not merged. Although the github API does not
describe this error we are assuming it is because github has not
calculated the mergeability of the PR in question. Thus the workaround
is to check mergeability and then try to merge again.