Process PRs oldest to newest
Today we process prs starting with the newest and working to the oldest.
This means that when the submit queue is empty a brand new PR might just
to the front of the line. Run PRs oldest to newest.

It actually might save us a couple of API calls, since the pages won't be
different on each loop if a new PR is entered. Only the last page will
change instead of every page. But that's pure speculation.