Get RateLimit from github on every call, not just once per loop
The code would collect how many tokens were left at the end of the loop
but when the tokens would reset at the beginning of the next loop. So if
we sleeped just before a token reset we would wake up thinking we didn't
have many tokens and that the next reset was an hour away.

This parses the github headers on every API response and updates the
timeout and remaining count. Thus we should always be correct.