Leave queue intact when e2e tests are failing
This means that you can always see the queue in the web interface, we'll
always merge in order, etc.

You'll also get a single message in the history when the queue
breaks/repairs, instead of a flood....

This also removes the whole 'wakeup' stuff where the worker thread
sleeps forever. Instead the worker thread will sleep for 30 seconds if
it can't do any work and then try again. Simpler code.