New munger to rerun CI if it is 'pending' for 24 hours
At times the CI can mess up, say it is testing a PR, but not actually be
testings.  This munger will request a retest when it finds a mergeable
PR with LGTM which has been 'pending' for 24 hours.

We should really retest every PR which is mergeable and has been
'pending' for greater than some length of time, but at least LGTM will
keep the jenkins queue from exploding...