Get e2e test results from builders via GCS
Federated builders are configured using a JSON file in the config
directory.  Each specifies a Google Cloud Storage bucket to crawl for
results in the known Jenkins format: latest-build.txt pointing to a
build result with finished.json.

This means that builders can be remote and not in the same network as
the munger; they need not even be Jenkins based!