Implement a new tool to watch for cherrypicks
This requires moving some things around as we now have one binary which
serves 2 very distinct web pages (submit-queue and cherrypick-queue)

Both queues need to look at very different sets of PRs. submit-queue
cares about open PRs. Cherrypick needs to look at open and closed PRs.
While at first glance it might seem easy to just make the submit-queue
ignore closed PRs (it is easy) it would mean that the submit queue would
have to walk all 22000+ issues on every loop, obviously ignoring most of
them.

Instead for the cherry-pick we are able to limit the issues we walk to a
given label (cherrypick-candidate), so even though we are walking all
issues (open or close) that number becomes quite managable. Obviously we
can't just walk the issues with that label for the submit queue...

Thus we end up with 1 binary which is becoming two distinct 'apps' with
distinct configuration. The makefile defaults to deploying and working
with the 'submit-queue' APP, but one can use APP=cherrypick