Adds Support for Kubernetes Internal Network Management
Previously, the Ansible project only supported managing a
single interface. It's common for production deployments to
require separating network traffic. This patch adds support
to manage an internal interface (no default gw) and associated
static routes.

1. Supports CoreOS deployments (systemd-networkd).
2. Adds static route support (including multiple routes).

The code can be used as a reference for supporting other
distros.

Note: This PR does not configure any service to specifically
bind to a particular interface.