Updates Vagrant README
1. Adds Ansible 2.0.0.2 for pip install
2. Updated NUM_MINIONS to NUM_NODES

The default version of Ansible for Mac OSX 10.10.5 causes errors
with coreos Ansible playbooks. THis commit hardcodes an Ansible
version that is compatible acorss distros. The Vagrantfile was
updated in a previous commit to go from NUM_MINIONS > NUM_NODES
but the README was not updated. This commit updates NUM_MINIONS
to NUM_NODES in the Vagrant readme.