Warn before closing stale pull-request
Create a warning mechanism so that we don't close pull-requests
overnight. You can choose how much time before you want to be warned,
and how often you want to be reminded.

This should also remove previous warning comments from the list of
comments so that we don't clutter the pull-request log.