Close inactive pull-request
Close a pull-request if it didn't have any human interaction in a given
time. It does so by going through all the comments of the
issue/pull-request not made by a bot and through events looking for a
re-open.

We can not find events such as a new push to the pull-request for the
moment.

The closing comment will be removed if the pull-request is re-opened