Make copy when deleting instead of messing up range operations
We would walk the labels and comments lists and could then delete from
them. Since we used the form `a = append(a[:i], a[i+1:...])` we were
actually changing the list as it was being used in a `range a`
statement. This could cause the comment deleter and the LGTM label
checker to go nuts.