Ignore invalid new resources
When github returns an invalid resource (comment, issue, event) we parse
it into the orm format. If a mandatory field is missing, the conversion
format will return nil. Currently, the return value is not checked and
the application crashes. Let's ignore invalid resources by moving to the
next when it happens.