Label doc only PRs with retest-not-required-docs-only
And modify the submit-queue to treat this label like it treats the
retest-not-required label.

The mungegithub will label pull-requests that don't have actual change
with retest-not-required-docs-only label. If the pull-request is
modified to add non-doc changes, it will remove the label. Because it
wouldn't make much sense to remove the retest-not-required label when
the pull-request changes (maybe someone added it manually for a good
reason), we have created this new label.