Create the Issue Label Triager
Create a machine learning tool to triage the github issue labeling
process.  Offline, we have used some validation techniques to select a
good model and hyperparameters (regularization parameters) for the ml
model.

Then, we use the flask microframework to expose the machine
learning model via a webapp that exposes two endpoints:
	/ [POST]: ask the ml model to predict labels for an issue's
		title and body
	/update_models [PUT]: train or update the ml model
We store the model in a 10GB persistent volume on GCP.

The deploy.sh file controls the gke deployment. It builds th
container, registers it (a cmdline param for the script), and exposes a
a service on port 5000 so that other pods can query for webapp for
labels.  It creates the deployment on gke and mounts the PV at /model/,
and it starts the webapp.