add playbook for kubernetes-addons role, decrease dependency from master to kubernetes
Introduce new playbook for kubernetes-addons role.

At the same time decrease dependency of the role from master to kubernetes.
With the master dependency everytime the addons is run (e.g. to update addons),
master role is run as well (resulting in update or re-configuration)
which is not always desired. On the other hand, if the addons role is run,
functioning master is required in advance.