Fix size check
The github api lies about the size of diffs if they are large. If we
then subtract out the size of generated files that it is not lieing
about we end up with negative numbers. So huge PRs can get XS size.

Instead of believing the API and deleting those things we don't care
about just sum the things we do care about and use that.