Update README.md
According to the docs page, "If a node is not needed for more than 10 min (configurable) then it can be deleted."
Where would I configure that 10 minute value. Since it will wait for 10 minutes per node it needs to delete, it could take a while to remove all the cruft if I have a lot of overage.

"I believe the value is not configurable." - David Oppenheimer from GCP.

If this value is configurable, please close this PR and document how to configure it. If it really isn't configurable, this is a missing feature.