mungegithub: Ignore missing ALIAS file
Ignore if the file is not present and can't be read. The mungebot
configuration should be able to say where the file is expected to be
even if it's not there yet.

Whenever the file fails to be read because it's not present, let's just
reset the current alias mapping to an empty mapping.