Delete cached github responses we know are bad
Github seems to screw up their conditional responses when the last entry
is exactly on a page size boundary. Today, when we get the last page
which is exactly full we try the next page anyway even though github
says there isn't a next page.

This patch attempts to delete our cached entry when we find that case.