Updates the TempFile Location
This commit updates the TempFile Location from SSL Certificates. This is necessary, as the os.Rename directive doesn't work on different mount points.

As this program is being used also on bare metal machines (and not only containers), the temp file should be created on the same directory of the final SSL certificates.