mungegithub: Stop assuming we'll receive each issues
The mungebot now only munges issues that have been modified since the
last loop. Some of the mungers made the assumptions that they would
receive all the issues, by refreshing their list for example. Try to
refresh the list in a better way so that we don't have to retrieve all
the list.