Allows certificate to be generated for multiple masters at once
Fixed the quotes in ansible filter

Removed task from another PR

Fixed the master name splitting

This should fix the problems with adding the masters to the san_array.

Finally found the cause of the problem why the script errors

All masters need to be in the alt subject list

Removed the fixme comment for cert gen
Thanks for noticing, @xialonglee!

cert_ip should be the IPs of all masters

"Remade" the kube_cert_ip stringy

Fixed the "fix" for a default var in Ansible

Change the name of the cert to be master

Removed quotes around uneccessary strings

Fix the cert_ip

Add the IPs using the IFS for loop

Create comma separated kube_cert_ip list