Allows certificate to be generated for multiple masters at once
Fixed the quotes in ansible filter

Removed task from another PR

Fixed the master name splitting

This should fix the problems with adding the masters to the san_array.

Finally found the cause of the problem why the script errors

All masters need to be in the alt subject list

Removed the fixme comment for cert gen
Thanks for noticing, @xialonglee!