ansible: add cluster_public_hostname
The cluster_hostname is already present (to be able to reach the
private/internal vip kube api).
The goal of cluster_public_hostname is to have the same thing but for
public access to the api.

If the public vip domain is not part of the certificat, when you will
try to access to the api via kubectl you will have something like that :

Unable to connect to the server: x509: certificate is valid for
cluster_hostname, master0, master1 ..., not cluster_public_hostname