Allowing to keep ca.key file
Add kube_cert_keep_ca variable in group_vars to be able to keep ca.key file.

This file can be usefull to be able to create user certificates, we need to have ca.key on one server (masters).

It's recommanded to remove that file from the server after generation and to keep it outside the cluster.

Fixes #2394
Fixes #651