Watch multiple components by one prometheus-to-sd.
Before the change, one instance of prometheus-to-sd was able to fetch
data from only one component.

Now, --source flag can be provided multiple times to describe multiple
components to be watched by one instance of prometheus-to-sd.

For backward compatibility, combination of --component, --host, --port
and --whitelisted-metrics is still supported as a method of describing
component to watch.