Update MetricDescriptor if label set has changed.
If newest version of metric descriptors has different set of label than
the original one, then update metric descriptor in the stackdriver. If
for some reason metric descriptor could not be updated, then we are
marking this metric as a broken one and corresponding timeserieses are
not going to be send.